sync: add NewSemaphoreKey for a unix semaphore.

+++ sync/sepmaphore_unix.go
@@ -16,19 +16,30 @@ type Semaphore struct {
 	id   int
 }
 
-// NewSemaphore creates a new sysV semaphore.
-//	name - object name
+// NewSemaphore creates a new sysV semaphore with the given name.
+// It generates a key from the name, and then calls NewSemaphoreKey.
+func NewSemaphore(name string, mode int, perm os.FileMode, initial int) (*Semaphore, error) {
+	k, err := common.KeyForName(name)
+	if err != nil {
+		return nil, err
+	}
+	result, err := NewSemaphoreKey(uint64(k), mode, perm, initial)
+	if err != nil {
+		return nil, err
+	}
+	result.name = name
+	return result, nil
+}
+
+// NewSemaphoreKey creates a new sysV semaphore for the given key.
+//	key - object key. each semaphore object is identifyed by a unique key.
 //	mode - object creation mode. must be one of the following:
 //		O_OPEN_OR_CREATE
 //		O_CREATE_ONLY
 //		O_OPEN_ONLY
 //	perm - object permissions
 //	initial - this value will be added to the semaphore's value, if it was created.
-func NewSemaphore(name string, mode int, perm os.FileMode, initial int) (*Semaphore, error) {
-	k, err := common.KeyForName(name)
-	if err != nil {
-		return nil, err
-	}
+func NewSemaphoreKey(key uint64, mode int, perm os.FileMode, initial int) (*Semaphore, error) {
 	var id int
 	creator := func(create bool) error {
 		var creatorErr error
@@ -36,14 +47,14 @@ func NewSemaphore(name string, mode int, perm os.FileMode, initial int) (*Semaph
 		if create {
 			flags |= common.IpcCreate | common.IpcExcl
 		}
-		id, creatorErr = semget(k, 1, flags)
+		id, creatorErr = semget(common.Key(key), 1, flags)
 		return creatorErr
 	}
 	created, err := common.OpenOrCreate(creator, mode)
 	if err != nil {
 		return nil, err
 	}
-	result := &Semaphore{name: name, id: id}
+	result := &Semaphore{id: id}
 	if created && initial > 0 {
 		if err = result.Add(initial); err != nil {
 			result.Destroy()
@@ -63,7 +74,7 @@ func (s *Semaphore) Add(value int) error {
 // Destroy removes the semaphore permanently.
 func (s *Semaphore) Destroy() error {
 	err := semctl(s.id, 0, common.IpcRmid)
-	if err == nil {
+	if err == nil && len(s.name) > 0 {
 		if err = os.Remove(common.TmpFilename(s.name)); os.IsNotExist(err) {
 			err = nil
 		}
